+++ API/Controllers/InstituteController.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services.Models;
 using Services.Models.Response;
@@ -7,6 +8,7 @@ namespace UrfuTestTask.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize]
 public class InstituteController : ControllerBase
 {
     private readonly IInstituteService _instituteService;
@@ -16,37 +18,50 @@ public class InstituteController : ControllerBase
         _instituteService = instituteService;
     }
 
+    // Получить все институты
     [HttpGet]
     public async Task<IActionResult> GetAllInstitutes()
     {
         var institutes = await _instituteService.GetAllInstitutesAsync();
         return Ok(institutes);
     }
-
+    
+    // Получить институт по Id
     [HttpGet("{id}")]
+    [ProducesResponseType(typeof(InstituteResp), 200)]
+    [ProducesResponseType(404)]
     public async Task<IActionResult> GetInstituteById(Guid id)
     {
         var institute = await _instituteService.GetInstituteByIdAsync(id);
         if (institute == null)
         {
-            return NotFound();
+            return NotFound("Institute not found");
         }
         return Ok(institute);
     }
-
-    [HttpPost]
     
-    [ProducesResponseType<InstituteResp>(200)]
+    // Добавить институт
+    [HttpPost]
+    [ProducesResponseType(typeof(InstituteResp), 200)]
     public async Task<IActionResult> AddInstitute([FromBody] InstituteReq instituteReq)
     { 
         var response = await _instituteService.AddInstituteAsync(instituteReq); 
         return Ok(response);
     }
-    [HttpDelete]
-    public async Task<IActionResult> DeleteModule(Guid id)
+    
+    // Удалить институт
+    [HttpDelete("{id}")]
+    [ProducesResponseType(200)]
+    [ProducesResponseType(404)]
+    public async Task<IActionResult> DeleteInstitute(Guid id)
     {
+        var existingInstitute = await _instituteService.GetInstituteByIdAsync(id);
+        if (existingInstitute == null)
+        {
+            return NotFound("Institute not found");
+        }
+
         await _instituteService.DeleteInstituteAsync(id);
         return Ok("Institute deleted successfully");
     }
-    
 }
